+++ app/Mail/EmailCode.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace App\Mail;
+
+use Illuminate\Bus\Queueable;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Mail\Mailable;
+use Illuminate\Queue\SerializesModels;
+
+class EmailCode extends Mailable
+{
+    use Queueable, SerializesModels;
+
+    public $code;
+
+
+    public function __construct($code)
+    {
+        $this->code = $code;
+    }
+
+    /**
+     * Build the message.
+     *
+     * @return $this
+     */
+    public function build()
+    {
+        return $this
+            ->from('info@3x1.io')
+            ->view('emails.code');
+    }
+}