+++ resources/views/emails/code.blade.php
@@ -1,0 +1,3 @@
+<div>
+    <h1 style="text-align: center">{{$code}}</h1>
+</div>