+++ resources/views/emails/reset.blade.php
@@ -1,0 +1,3 @@
+<div>
+    <h1 style="text-align: center">Your Password Has Been Reset Success</h1>
+</div>