+++ resources/views/emails/welcome.blade.php
@@ -1,0 +1,3 @@
+<div>
+    <h1 style="text-align: center">Welcome To Our App</h1>
+</div>