+++ CHANGELOG.md
@@ -4,3 +4,9 @@
 - Fix Database migration
 - Fix Main Theme
 - Add Blocks Feature
+
+# v1.02
+- Add Auth APIs
+- Add User CLI APIs
+- User Activated By Email
+- User Activated By Phone [MessageBird]